[#228][be][schedule]refactor: 누락된 코드 추가
